+++ src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/compute_waves/fetch_proc.F90
@@ -35,8 +35,8 @@ module fetch_proc_operation_data
 
    integer, dimension(:), allocatable :: ndx_over_procs !< ndx data over procs
    integer, dimension(:, :), allocatable :: iglobal_s_procs !< iglobal_s data for each proc
-   double precision, dimension(:, :), allocatable :: f_buffer !< buffer to send fetch/fetdp data
-   double precision, dimension(:), allocatable :: s1_buffer !< buffer to send s1 data
+   real(kind=dp), dimension(:, :), allocatable :: f_buffer !< buffer to send fetch/fetdp data
+   real(kind=dp), dimension(:), allocatable :: s1_buffer !< buffer to send s1 data
 
    integer :: dflowfm_entire_group !< this group includes all procs
    integer :: dflowfm_group !< this group includes all procs except the last one which is the fetch proc
@@ -66,9 +66,9 @@ integer function initialise_fetch_proc_data() result(iresult)
    integer :: ndx_max, index_loop, error, source, tag = 100, icount
    integer, dimension(:), allocatable :: iglobal_s_source
    integer :: flow_node, isearch, j0
-   double precision, dimension(:), allocatable :: xz_proc, yz_proc
+   real(kind=dp), dimension(:), allocatable :: xz_proc, yz_proc
    logical :: iglobal_s_exist_on_fetch_proc = .false.
-   double precision, parameter :: tolerance = 1.0d-3
+   real(kind=dp), parameter :: tolerance = 1.0d-3
 
 #ifdef HAVE_MPI
    if (allocated(ndx_over_procs)) deallocate (ndx_over_procs)
