+++ src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/prepost/updatevaluesonlaterals.f90
@@ -41,11 +41,11 @@ subroutine updateValuesOnLaterals(tim1, timestep)
    use m_flowparameters, only: eps10
    use m_partitioninfo, only: jampi, reduce_double_sum, is_ghost_node
    implicit none
-   double precision, intent(in) :: tim1 !< Current (new) time
-   double precision, intent(in) :: timestep !< Timestep is the difference between tim1 and the last update time
+   real(kind=dp), intent(in) :: tim1 !< Current (new) time
+   real(kind=dp), intent(in) :: timestep !< Timestep is the difference between tim1 and the last update time
 
    integer :: k1, i, i_lat, i_layer, i_node
-   double precision, allocatable :: qLatRealCumTmp(:), qLatRealMPI(:)
+   real(kind=dp), allocatable :: qLatRealCumTmp(:), qLatRealMPI(:)
 
    ! If current time has not reached the history output start time yet, do not update
    if (comparereal(tim1, ti_hiss, eps10) < 0) then
