+++ src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_utils/solve_guus.F90
@@ -506,7 +506,7 @@
     implicit none
 
     integer :: ndx
-    double precision :: s1(ndx)
+    real(kind=dp) :: s1(ndx)
     integer :: itsol
     integer :: ierror ! error (1) or not (0)
 
@@ -627,8 +627,8 @@
     
     implicit none
     integer :: ndx, its
-    double precision :: s1(ndx)
-    double precision, dimension(Ndx), intent(in) :: righthandside !< right-hand side, all cells
+    real(kind=dp) :: s1(ndx)
+    real(kind=dp), dimension(Ndx), intent(in) :: righthandside !< right-hand side, all cells
     integer, intent(in) :: jaini !< compute preconditioner and permutation (1) or not (0), or initialization only (-1), or ILU solve only (2)
     integer, intent(in) :: jadosafe !< thread-safe (1) or not (0), will set jasafe module variable
     integer, intent(out) :: ierror !< error (1) or not (0)
@@ -636,7 +636,7 @@
     integer :: i, j, jj, n, ntot, na, nietnul, m
     integer :: minp, k
 
-    double precision :: res
+    real(kind=dp) :: res
 
     logical, save :: jaoutput = .false.
 
@@ -949,11 +949,11 @@
 
     implicit none
     integer :: ndx, ipre
-    double precision :: s1(ndx)
+    real(kind=dp) :: s1(ndx)
 
     integer :: i, j, jj, n, ntot
-    double precision :: rkzki, rkzki0, pkapki, alfak, betak
-    double precision :: eps
+    real(kind=dp) :: rkzki, rkzki0, pkapki, alfak, betak
+    real(kind=dp) :: eps
 
 ! ddr (rechterlid), bbr (diag) , ccr (off diag), s1, row, row()%j, row()%a [AvD]
     if (nocg <= 0) return
@@ -1096,11 +1096,11 @@
 
     implicit none
     integer :: ndx
-    double precision :: s1(ndx)
+    real(kind=dp) :: s1(ndx)
 
     integer i, ipre, j, jj, n
-    double precision :: rkzki, rkzki0, pkapki, alfak, betak
-    double precision :: eps
+    real(kind=dp) :: rkzki, rkzki0, pkapki, alfak, betak
+    real(kind=dp) :: eps
 
     if (nocg <= 0) return
 
@@ -1257,11 +1257,11 @@
 
     implicit none
     integer :: ndx
-    double precision :: s1(ndx)
+    real(kind=dp) :: s1(ndx)
 
     integer i, ipre, j, jj, n
-    double precision :: rkzki, rkzki0, pkapki, alfak, betak
-    double precision :: eps
+    real(kind=dp) :: rkzki, rkzki0, pkapki, alfak, betak
+    real(kind=dp) :: eps
 
     if (nocg <= 0) return
 
@@ -1411,7 +1411,7 @@
     implicit none
 
     integer :: ndx
-    double precision :: s1(ndx)
+    real(kind=dp) :: s1(ndx)
 
     integer m, n, np
     do n = nogauss, 1, -1
@@ -1431,7 +1431,7 @@
     implicit none
 
     integer :: ndx
-    double precision :: s1(ndx)
+    real(kind=dp) :: s1(ndx)
 
     integer m, n
     do n = nogauss, 1, -1
@@ -1541,7 +1541,7 @@
  subroutine gauss_elimination
     use m_reduce
     implicit none
-    !double precision :: ccc(500)
+    !real(kind=dp) :: ccc(500)
     integer j, k, m, n, np, m1, nodm1, m1m2
     ccc = 0d0
     do n = 1, nogauss
@@ -1572,7 +1572,7 @@
  subroutine gauss_eliminationjipjan
     use m_reduce
     implicit none
-    !double precision :: ccc(500)
+    !real(kind=dp) :: ccc(500)
     integer j, k, m, n, m1, nodm1, m1m2, mm, jj
 
     ccc = 0d0
@@ -1872,7 +1872,7 @@
 
     implicit none
     integer :: ndx, ipre
-    double precision :: s1(ndx)
+    real(kind=dp) :: s1(ndx)
     integer, intent(out) :: nocgiter_loc
     integer, intent(out) :: ierror !< error (1) or not (0)
     character(len=100) :: message
@@ -1881,11 +1881,11 @@
     integer :: nopreconditioner
 
     integer :: i, j, jj, n, ntot
-    double precision :: rkzki, rkzki0, pkapki, alfak, betak
-    double precision :: eps
+    real(kind=dp) :: rkzki, rkzki0, pkapki, alfak, betak
+    real(kind=dp) :: eps
 
     ! BEGIN MPI
-    double precision :: eps_tmp, rkzki_tmp, pkapki_tmp
+    real(kind=dp) :: eps_tmp, rkzki_tmp, pkapki_tmp
     ! END MPI
 
     ierror = 0
@@ -2384,18 +2384,18 @@
     implicit none
 
     integer, intent(in) :: ndx
-    double precision, dimension(Ndx), intent(inout) :: s1
+    real(kind=dp), dimension(Ndx), intent(inout) :: s1
     integer, intent(out) :: itsol
     integer, intent(out) :: ierror
 
     character(len=128) :: message
 
-!    double precision, dimension(:), allocatable      :: bdum, cdum, ddum
+!    real(kind=dp), dimension(:), allocatable      :: bdum, cdum, ddum
 
-    double precision :: maxdiff
-    double precision :: res ! residual
-    double precision :: dum
-    double precision :: beta, val
+    real(kind=dp) :: maxdiff
+    real(kind=dp) :: res ! residual
+    real(kind=dp) :: dum
+    real(kind=dp) :: beta, val
     
     integer, parameter :: MAXITER = 100
     integer :: iter, its
@@ -2405,9 +2405,9 @@
 
     integer, parameter :: javerbose = 1
 
-    double precision, dimension(:), allocatable :: bbr_sav, ddr_sav
-    double precision, dimension(1) :: res_global
-    double precision :: res_global0, tolDD
+    real(kind=dp), dimension(:), allocatable :: bbr_sav, ddr_sav
+    real(kind=dp), dimension(1) :: res_global
+    real(kind=dp) :: res_global0, tolDD
     if (nocg <= 0) return
 
     ierror = 0
